structural changes to code to merge explore tools
